+++ hawkey/tests/fixtures.c
@@ -190,11 +190,11 @@ void setup_yum_sack(HySack sack, const char *yum_repo_name)
     fail_if(access(repo_path, X_OK));
     HyRepo repo = glob_for_repofiles(pool, yum_repo_name, repo_path);
 
-    fail_if(hy_sack_load_yum_repo(sack, repo,
-				  HY_BUILD_CACHE |
-				  HY_LOAD_FILELISTS |
-				  HY_LOAD_UPDATEINFO |
-				  HY_LOAD_PRESTO));
+    fail_if(hy_sack_load_repo(sack, repo,
+                              HY_BUILD_CACHE |
+                              HY_LOAD_FILELISTS |
+                              HY_LOAD_UPDATEINFO |
+                              HY_LOAD_PRESTO));
     fail_unless(hy_sack_count(sack) == TEST_EXPECT_YUM_NSOLVABLES);
     hy_repo_free(repo);
 }