+++ hawkey/tests/python/tests/test_advisoryref.py
@@ -45,7 +45,7 @@ class Test(base.TestCase):
     def setUp(self):
         """Prepare the test fixture."""
         sack = base.TestSack(repo_dir=self.repo_dir)
-        sack.load_yum_repo(load_updateinfo=True)
+        sack.load_repo(load_updateinfo=True)
         self.reference = find_reference(
             sack, 'https://bugzilla.redhat.com/show_bug.cgi?id=472090')